Use tcsetattr & friends and O_NONBLOCK
This commit changes the serial initialization routine so that the POSIX
tcsetattr functions and friends are used instead of Linux specific termios2
ioctls.  Also, the device is opened with O_NONBLOCK during initialization
so that the CLOCAL can be set.  This prevents the initialization code from
hanging if the default modes require CD to be present on the serial port
(i.e. ~CLOCAL mode).